Add restAPI.py
This module currently takes 3 api request:
GET: get-fill-price
GET: get-currency-balance
POST: send-order

The functions check for errors in the requests themselves and return success if the requests are properly formatted.
The functions check that the arguments are valid (ie. valid markets/currencies) and checks that there are only the required amount of arguments.